+++ app/src/main/java/ua/com/radiokot/photoprism/features/gallery/view/model/GalleryViewModel.kt
@@ -535,6 +535,17 @@ class GalleryViewModel(
             is State.Selecting -> {
                 if (state.allowMultiple) {
                     toggleMediaMultipleSelection(media)
+
+                    // If the last selected item has been unselected,
+                    // automatically switch back to viewing.
+                    // But only if was selecting for user.
+                    if (state is State.Selecting.ForUser
+                        && multipleSelectionFilesByMediaUid.isEmpty()
+                    ) {
+                        log.debug { "onItemClicked(): unselected_last_switching_to_viewing" }
+
+                        switchToViewing()
+                    }
                 } else {
                     selectMedia(media)
                 }