+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support of Apple live photos uploaded as HEIC+MOV bundle
 - Ability to download gallery items selected with a long press
+- Exit the selection mode when unselecting the last selected gallery item
 
 ## [1.19.0] - 2023-09-24